+++ android/app/src/main/java/com/now/naaga/data/remote/dto/RankDto.kt
@@ -1,4 +1,4 @@
-package com.now.naaga.data.remote.dto.third
+package com.now.naaga.data.remote.dto
 
 import kotlinx.serialization.Serializable
